Moved ShoppingList
ShoppingList component is now in a more appropriate location
