+++ packages/ui/src/components/shopping/index.js
@@ -1,0 +1,1 @@
+export * from './ShoppingList';
